examples/everscale-sample/requirements.txt
ton-client-py==1.32.0.0
